Refactor Installation Enrichment (#53)
* Rename InstallationEnricher to InstallationVisitor

Enricher is a pretty generic term, Visitor is a more familiar name for
Java developers. Normally, the visitor pattern would have both a
`Visitable` and `Visitor` interface, enumerating each type of object
that could be visited. This is also known as double dispatch. However,
because there is only one type of object that can be visited we do not
need the additional weight that would usually be associated with this
pattern. Arguably, this disqualifies what our scenario from technically
being a visitor. However, we feel the value of using a well known type
name outweighs that argument.

* Removing Middleware concept in favor of just the InstallationVisitor

* Re-ignore tests

* Fix TagEnricherTest

* Rename TagEnricherTest to TagVisitorTest