Revert "Remove all references to dynamic resources (#66)" (#78)
This reverts commit 4fea1078ae63f84d9256ad5f42c2ebfe7a0b7cb0.

We originally moved this string to be a constant defined in code in response to distribution trouble. We weren't able to reference any Android specific resources, and #66 was a knee-jerk reaction. We've since come to understand that our maven plugin wasn't building the manifest correctly, so only the JAR was being used. We are now able to distribute everything that lives in an AAR, so it is safe to move back to a world where we use the strings file.