Removing Duplicate Network Status Code
Formerly, there was a Network Status broadcast receiver for handling
just requesting new Installations. When we updated the HttpClient, which
has its own version of Network availability code, we didn't catch the
duplication. Now that we're merging the HttpClient code into the
existing namespace, it didn't make sense to keep both any longer. The
version associated with the HttpClient appeared to be more robust, so
we chose to continue with that implmentation.