Adding InstallationAdapter and InstallationVisitor
More stashed progress

Incremental progress stashed

Adding ExpirationVisitor and ExpirationAdapter.

These classes will work together to ensure that by the time we're about
to call Notification Hubs' backend, a reasonable expiration date is
applied to the Installation.

ExpirationVisitor caches the most recent expiration date that was used,
and reapplies it when building future Installations. This is necessary
to have a deterministic Installation.

ExpirationAdapter considers whether or not the final Installation needs
to have its expiration updated, and informs the ExpirationVisitor if it
does. This protects the backend, and is able to decide whether or not an
Installation refresh is warranted or not.