Support Installation Expiration (#85)
* Adding InstallationAdapter and InstallationVisitor

More stashed progress

Incremental progress stashed

Adding ExpirationVisitor and ExpirationAdapter.

These classes will work together to ensure that by the time we're about
to call Notification Hubs' backend, a reasonable expiration date is
applied to the Installation.

ExpirationVisitor caches the most recent expiration date that was used,
and reapplies it when building future Installations. This is necessary
to have a deterministic Installation.

ExpirationAdapter considers whether or not the final Installation needs
to have its expiration updated, and informs the ExpirationVisitor if it
does. This protects the backend, and is able to decide whether or not an
Installation refresh is warranted or not.

* Plumbing through Expiration get/set ability

* Fixing equality comparison bug.

* Parameterize duration of validWindow and nearExpiration

* Replacing ExpirationAdapter/ExpirationVisitor with method in client

This wildly simplifies the process, but at the expense of reusability.
It's a trade-off we're willing to accept for now.

* Update debouncer to consider Installation stale after a set window

This will allow for an opprotunity to update the expiration at least
once a day. Additionally, if there's a fluke where the Debouncer is in
the way of updating the Installation, this allows this situation an
attempt at remdiying itself.