New api version (#82)
* Update Android SDK to use new API version

* use new version everywhere

* Simplify API Version switch

Someday we'll need the additional complexity around API-Version, and it
may look very similar to the previous  approach! For now we've papered over it
server side. I suspect that when we do introduce the additional
complexity, it'll come with full separate clients similar to the way the
AutoRest generates clients.

Co-authored-by: Martin Strobel <marstr@microsoft.com>