Rename Sample Applications (#93)
* notification-hubs-test-app -> notification-hubs-test-app-legacy

This change is the first step in indicating that the new namespace is
the preferred model, and the one that should be favored moving forward.

Note, this does not amount to a deprecation of the old namespace.

* notification-hubs-test-app-refresh -> notification-hubs-test-app-java

The updated test application takes the suffix "-java" in order to allow
room for a future clone of this application, but written in Kotlin.

* Update CI to target updated applications.

* Update legacy app title