Remove test that is causing flakiness.
This test was in place to ensure that the debouncer would stash data
in persisted memory. The only way for us to do that is to white-box test
it and look into persisted memory after the bulk of the test is
finished. However, that's causing flakiness. My current guess is that
while we have a lot of control over what happens during the test, that
our Network availability controller, or whatever else, is triggering a
new installation to be put, and suddenly our cache gets changed
underneath us in the middle of a test.

Personally, I don't think the test provides much in the way of
regression protection as we are unlikely to accidently stop storing in
SharedPreferences. Also, it's low stakes if we did, as that would just
mean more installation traffic is sent to the server. For that reason,
instead of trying to fix this test, I'm just removing it.