Fix CI Build (#102)
* Remove test that is causing flakiness.

This test was in place to ensure that the debouncer would stash data
in persisted memory. The only way for us to do that is to white-box test
it and look into persisted memory after the bulk of the test is
finished. However, that's causing flakiness. My current guess is that
while we have a lot of control over what happens during the test, that
our Network availability controller, or whatever else, is triggering a
new installation to be put, and suddenly our cache gets changed
underneath us in the middle of a test.

Personally, I don't think the test provides much in the way of
regression protection as we are unlikely to accidently stop storing in
SharedPreferences. Also, it's low stakes if we did, as that would just
mean more installation traffic is sent to the server. For that reason,
instead of trying to fix this test, I'm just removing it.

* Disable NotificationHub runs during tests

* Revert "Disable NotificationHub runs during tests"

This reverts commit 441d3ec1787b4a5b7d9d72c73760039c5f51d1bc.

* Remove unused imports

* Overhaul debouncer tests, and change timing for accepting installation.

* Revert "Remove test that is causing flakiness."

This reverts commit b2561b5bbbdbd96cdfa12996532deaaf930b2bac.

It also modifies the test previously deleted to match overhauled
format.

* Optimize imports

* Use MockSharedPreferences for TagVisitorTest as well.

* Move overhauled tests out of androidConnectedTests

This will make them faster and plug into unit-testing frameworks much
more easily.

* Move TemplateVisitorTests off of device.

* Move InstallationTemplateTest off device

* Optimize Imports

* Move legacy tests off of device.

* Remove HttpClientRetryerTests

I plan on removing the HTTP client that was copied over from AppCenter
anyway. By deleting this test, I'm able to remove the whole step of
installing an Android emulator from CI.

* Remove emulator from CI.

* Revert "Remove emulator from CI."

This reverts commit 3d474429d241c07f526719845d77d084b109a6b9.

* Remove unused section to CI

* Remove mockito