Don't allow thread interrupt when cancelling InstallationPut
Because we don't have the separation of a thread pool delivering
responses vs a thread pool executing requests like Volley does, we could
get ourselves in a position where we cancel a user's ability to react to
an error in the middle of their error handler. We don't want that.