Fixing Firebase initialization race condition.
The NotificationHub global single instance needs to have a reference to
the Application in order to load SharedPreferences. Before this change,
the only time we could get access to that Application was through the
User's code calling us from an Activity or Service. In some situations,
like the one described in #123, FirebaseReceiver was being initialized
before it made sense for the customer to call NotificationHub.start().
In that case, NotificationHub was uninitialized, and FirebaseReceiver
crashes with a NullReferenceException, inhibiting further usage. By
refactoring the actual visitor population out, it allows either
FirebaseReceiver or the customer to win the race and initialize
NotificationHub.