Tuning the Default Max Retry Count.
We came up with this number empirically using the sample application,
after finding that it consistently takes 5 passes through the
Installation pipeline to have a PushChannel assigned. We decided that 15
operations without a push channel assignment felt more likely to trip
than 95, but still not often enough to be noisy.