Read Version from BuildConfig class (#160)
* Read version from BuildConfig class.

This cleans up a long standing frustration with hadrd coding this version number. In addition, it allows us to have separate version strings based on the SKU (FCM, ADM, or Baidu optimized moving forward.)

* Don't process Javadoc for InstallationPutRequest.java

This prevents a failure to generate Javadoc because it can't find the generated source for the BuildConfig type.