Be more aggressive about fetching the Firebase Token on application start
Previously, if we already had a token in-hand we rely on the onNewToken callback to let us know if it changed. However, to be more defensive, we will now always ask.

The layer that will prevent an installation update if the token matches the previous record will remain in place.