Make PushChannel bootstrapping code available
Previously, we only fetched the FirebaseMessaging token in the
FirebaseReceiver. This lead to a bug where bootstrapping didn't occur
during app-upgrades where new tokens are not issued, hence the
FirebaseReceiver is never instantiated.

With this change, the FirebaseMessaging ID will be re-evaluated when the
NotificatonHub.start method is called, and/or when the FirebaseReceiver
is started.